version() for key, added

+++ Key.h
@@ -58,7 +58,7 @@ namespace OpenPGP {
             std::string fingerprint() const;
 
             // version of entire key (primary key packet)
-            int version() const;
+            uint8_t version() const;
 
             // output style inspired by gpg and SKS Keyserver/pgp.mit.edu
             std::string list_keys(const std::size_t indents = 0, const std::size_t indent_size = 4) const;
